+++ lib/ooxml_parser/docx_parser.rb
@@ -10,8 +10,8 @@ module OoxmlParser
     # @param path_to_file [String] file path
     # @return [DocumentStructure] result of parse
     def self.parse_docx(path_to_file)
-      Parser.parse_format(path_to_file) do
-        DocumentStructure.new.parse
+      Parser.parse_format(path_to_file) do |path_to_folder|
+        DocumentStructure.new(unpacked_folder: path_to_folder).parse
       end
     end
   end