+++ lib/ooxml_parser/docx_parser/document_structure/styles.rb
@@ -17,7 +17,7 @@ module OoxmlParser
     # Parse styles data
     # @return [Styles] result of parsing
     def parse
-      doc = parse_xml("#{OOXMLDocumentObject.path_to_folder}word/styles.xml")
+      doc = parse_xml("#{root_object.unpacked_folder}word/styles.xml")
       doc.xpath('w:styles/*').each do |node_child|
         case node_child.name
         when 'docDefaults'