+++ lib/ooxml_parser/pptx_parser/presentation/slide.rb
@@ -51,9 +51,9 @@ module OoxmlParser
     # Parse Slide object
     # @return [Slide] result of parsing
     def parse
-      OOXMLDocumentObject.add_to_xmls_stack(@xml_path)
+      root_object.add_to_xmls_stack(@xml_path)
       @name = File.basename(@xml_path, '.*')
-      node = parse_xml(OOXMLDocumentObject.current_xml)
+      node = parse_xml(root_object.current_xml)
       node.xpath('//p:sld/*').each do |node_child|
         case node_child.name
         when 'cSld'
@@ -66,8 +66,8 @@ module OoxmlParser
           @alternate_content = PresentationAlternateContent.new(parent: self).parse(node_child)
         end
       end
-      OOXMLDocumentObject.xmls_stack.pop
-      @relationships = Relationships.new(parent: self).parse_file("#{OOXMLDocumentObject.path_to_folder}#{File.dirname(@xml_path)}/_rels/#{@name}.xml.rels")
+      root_object.xmls_stack.pop
+      @relationships = Relationships.new(parent: self).parse_file("#{root_object.unpacked_folder}#{File.dirname(@xml_path)}/_rels/#{@name}.xml.rels")
       parse_note
       self
     end
@@ -79,7 +79,7 @@ module OoxmlParser
       notes_target = @relationships.target_by_type('notes')
       return nil if notes_target.empty?
 
-      @note = PresentationNotes.new(parent: self).parse("#{OOXMLDocumentObject.path_to_folder}#{File.dirname(@xml_path)}/#{notes_target.first}")
+      @note = PresentationNotes.new(parent: self).parse("#{root_object.unpacked_folder}#{File.dirname(@xml_path)}/#{notes_target.first}")
     end
   end
 end