+++ lib/ooxml_parser/pptx_parser/presentation/slide/graphic_frame/graphic_frame.rb
@@ -30,9 +30,9 @@ module OoxmlParser
                 when 'tbl'
                   graphic_data << Table.new(parent: self).parse(graphic_node_child)
                 when 'chart'
-                  OOXMLDocumentObject.add_to_xmls_stack(OOXMLDocumentObject.get_link_from_rels(graphic_node_child.attribute('id').value))
+                  root_object.add_to_xmls_stack(root_object.get_link_from_rels(graphic_node_child.attribute('id').value))
                   graphic_data << Chart.new(parent: self).parse
-                  OOXMLDocumentObject.xmls_stack.pop
+                  root_object.xmls_stack.pop
                 when 'oleObj'
                   graphic_data << OleObject.new(parent: self).parse(graphic_node_child)
                 end