+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 
 * `PresentationTheme#parse` now an instance method
+* Remove most of class vars from `OOXMLDocumentObject`
 
 ## 0.29.0 (2022-09-05)